articles/examples/langgraph/requirements.txt
# AgentCore LangGraph Example
# https://docs.aws.amazon.com/bedrock-agentcore/

boto3>=1.34.0
langgraph>=0.2.0
langchain-aws>=0.2.0
bedrock-agentcore-starter-toolkit>=1.0.0
python-dotenv>=1.0.0
